requirements.txt
faiss-cpu 
langchain 
newspaper3k 
google-api-python-client 
lxml-html-clean 
langchain-groq
python-dotenv
langchain-community
sentence-transformers
pypdf
streamlit
faiss-cpu
validators